Started implementing notifications actions
This commit start works on actionable notifications for the DawnPlus app. This change includes creating a notification category and and also the creation of sample sounds for notifications. the change also adds a few background execution modes to the app for improve control over the UX experience and conforming to user expectation.